requirements.txt
python-dotenv
pymongo
python-telegram-bot == 20.5
python-telegram-bot[job-queue]